refactor: centralize proxy config in httpClient.js
Move global proxy setup to httpClient.js as single source of truth,
removing duplicate code from configApi.js and gemini-client.js.
Import httpClient early in server.ts to ensure proxy is configured
before other modules use undici.
